openwrt: arc - remove dependency on DEVTMPFS
OpenWRT builds initramfs so that it doesn't require DEVTMPFS so dropping
this dependency. That helps to escape 2 separate kernel rebuilds with
and without initramfs.

2 builds happen because OpenWRT first builds kernel and later modules.
When building entire kernel with simple "make" INITRAMFS sets to a real
value and so was triggering DEVTMPFS selection. Then when building only
modules with "make modules" command INITRAMFS is zeroed and so kernel
config was changing that lead to full kernel rebuild.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>